351003/Naskevich/pyproject.toml
[project]
name = "distcomp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "cassandra-driver (>=3.29.3,<4.0.0)",
    "faststream[kafka] (>=0.6.7,<0.7.0)",
    "redis (>=5,<6)",
    "pyjwt[crypto] (>=2.12.1,<3.0.0)",
    "bcrypt (>=4.0.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
